SolnAI-agents/shared/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solnai-agent-utils"
version = "1.0.0"
description = "Shared utilities for SolnAI agents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SolnAI Team"}
]
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "pydantic>=2.0.0",
    "supabase>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.25.0",
    "httpx>=0.23.0",
    "openai>=0.27.0",
    "pytest>=6.2.5",
    "pytest-asyncio>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pytest>=6.2.5",
    "pytest-asyncio>=0.20.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
